CustomCommandsSQL: if the SQL version <= 3, then no need to reset the SQL to default when users install a new updated version of nh app.

+++ src/com/offsec/nethunter/SQL/CustomCommandsSQL.java
@@ -80,7 +80,9 @@ public class CustomCommandsSQL extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+        if (oldVersion > 3) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+        }
         this.onCreate(db);
     }
